[android] Added chart integration
