+++ android/res/layout-land/elevation_profile_bottom_sheet.xml
@@ -5,8 +5,6 @@
   android:id="@+id/elevation_profile"
   android:layout_width="@dimen/place_page_width"
   android:layout_height="wrap_content"
-  android:paddingStart="@dimen/margin_base"
-  android:paddingEnd="@dimen/margin_base"
   android:paddingTop="@dimen/margin_base"
   android:paddingBottom="@dimen/margin_base"
   android:background="?cardBackground"
