+++ android/res/layout/current_location_marker.xml
@@ -21,8 +21,6 @@
     android:text="you"
     android:paddingStart="@dimen/margin_quarter"
     android:paddingEnd="@dimen/margin_quarter"
-    android:layout_marginTop="@dimen/margin_quarter"
-    android:layout_marginBottom="@dimen/margin_quarter"
     android:background="@drawable/bg_your_location_pin"
     android:textSize="@dimen/text_size_icon_title"
     android:textColor="@color/base_accent"
