--- android/res/layout/elevation_profile_internal_layout.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:tools="http://schemas.android.com/tools"
-  tools:showIn="@layout/elevation_profile_bottom_sheet">
-  <LinearLayout
-    android:id="@+id/pull_icon_container"
-    android:layout_width="wrap_content"
-    android:layout_height="@dimen/margin_half"
-    android:layout_marginBottom="@dimen/margin_eighth"
-    android:layout_gravity="center_horizontal"
-    android:gravity="center"
-    android:orientation="vertical">
-    <ImageView
-      android:id="@+id/pull_icon"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content" />
-  </LinearLayout>
-  <TextView
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:paddingTop="@dimen/margin_half"
-    android:paddingBottom="@dimen/margin_base"
-    android:textAppearance="?fontHeadline6"
-    tools:text="Xindian Shitoushan Trail" />
-  <include layout="@layout/list_divider" />
-  <LinearLayout
-    android:id="@+id/chart_container"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:minHeight="143dp"
-    android:gravity="center"
-    android:orientation="vertical">
-    <TextView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="Chart is coming soon!" />
-  </LinearLayout>
-</merge>
